Added files via upload
The attached package was consolidated and created by buckeyestudios.com.  
"a package containing all the files you made for the fps weapons project so far I thought you might want to add it to a download for a tutorial you were making for this. This way you could have a package that was up to date with everything then a separate file with the settings for the weapons"